Adding reused and adapted single-core list proofs.
This is Nathan's list predicates and lemmas modified as required for the context switching proof. Files are renamed to avoid conflicts. For reference the files contain both Nathan's unmodified version and the modified version. To check which parts have been adapted, see instructions in README section "Comparing the Original List Proofs and Our Adaptation".